## Description

This PR introduces performance benchmark tests for SharedTree's table-related APIs, focusing on standard operations. The benchmarks cover:

Insertion of rows, columns, and cells

Includes insertion at the middle of the table for rows and columns. Similarly, tests removal for rows and columns.

## Test Result
~~~
/home/tong/workspace/FluidFramework/node_modules/.bin/mocha ./tableTree.spec.js --no-timeouts --exit
Writing test results relative to package to nyc/junit-report.xml
mocharc-common.cjs:73
Warning: Cannot find any files matching pattern "lib/test"
collect-files.js:102

spec.js:54
  SharedTree memory usage
spec.js:54
    Size of 10*10 SharedTree
spec.js:54
      Column Insertion
spec.js:54
        ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 5 times (149ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Undo insert column in the middle 5 times (96ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Redo insert column in the middle 5 times (89ms)
spec.js:83
      Column Insertion
spec.js:54
        ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 5 times (77ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Undo insert column in the middle 5 times (78ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Redo insert column in the middle 5 times (82ms)
spec.js:83
      Column and Row Insertion
spec.js:54
        ✔ @Benchmark @Measurement @MemoryUsage Insert a column and a row in the middle 5 times (74ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Undo insert column and row in the middle 5 times (73ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Redo insert column and row in the middle 5 times (77ms)
spec.js:83
      Column Removal
spec.js:54
        ✔ @Benchmark @Measurement @MemoryUsage Remove a column in the middle 5 times (88ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Undo remove column in the middle 5 times (96ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Redo remove column in the middle 5 times (91ms)
spec.js:83
      Row Removal
spec.js:54
        ✔ @Benchmark @Measurement @MemoryUsage Remove a row in the middle 5 times (61ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Undo remove row in the middle 5 times (63ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Redo remove row in the middle 5 times (63ms)
spec.js:83
      Column and Row Removal
spec.js:54
        ✔ @Benchmark @Measurement @MemoryUsage Remove a column and a row in the middle 5 times (77ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Undo remove column and row in the middle 5 times (78ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Redo remove column and row in the middle 5 times (78ms)
spec.js:83
      Insert a column and a row and remove right away
spec.js:54
        ✔ @Benchmark @Measurement @MemoryUsage Insert a column and a row in the middle and remove right away 5 times (77ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Undo insert and remove column and row in the middle 5 times (85ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Redo insert and remove column and row in the middle 5 times (83ms)
spec.js:83
      Cell Value Setting
spec.js:54
        ✔ @Benchmark @Measurement @MemoryUsage Set cell value in the middle 5 times (54ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Undo set cell value in the middle 5 times (59ms)
spec.js:83
        ✔ @Benchmark @Measurement @MemoryUsage Redo set cell value in the middle 5 times (95ms)
spec.js:83
  24 passing (2s)
~~~